1.支持对静态资源的http 206 范围（部分）请求,固定支持bytes级别文件断点续传 2.支持MP3/MP4文件类型 3.实现WebSocket层自动注入

+++ MyBatisX/src/pers/pandora/core/PoolConnection.java
@@ -1,23 +1,19 @@
 package pers.pandora.core;
 
 import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-
 public class PoolConnection {
 
     private boolean isBusy;
 
     private Connection connection;
 
-    private int transNew = -1;
+    private byte transNew = -1;
 
-    public int getTransNew() {
+    public byte getTransNew() {
         return transNew;
     }
 
-    public void setTransNew(int transNew) {
+    public void setTransNew(byte transNew) {
         this.transNew = transNew;
     }
